Fix warnings

+++ FluxEngine/Rendering/Model.cpp
@@ -50,7 +50,7 @@ void Model::SetMaterial(Material* pMaterial)
 
 void Model::SetMaterial(int index, Material* pMaterial)
 {
-	if (index >= m_Batches.size())
+	if (index >= (int)m_Batches.size())
 	{
 		FLUX_LOG(ERROR, "[Model::SetMaterial] > Index out of range! Is '%i' but model only has '%i' batches", index, m_Batches.size());
 		return;
